% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{estimate_pi2}
\alias{estimate_pi2}
\title{Estimation of Pi}
\usage{
estimate_pi2(B = 5000, seed = 10)
}
\arguments{
\item{B}{the number of points on which the estimation of pi is produced.}

\item{seed}{the seed to be used for the random generation used in the procedure}
}
\value{
A \code{list} of the `pi` class made of the estimated value of pi and the points on which this estimation is based. In particular:
\describe{
    \item{estimated_pi}{The estimated value of pi}
    \item{points}{A data frame with the points in the unit square on which the estimation is based and a dummy indicating whether the particular point is inside the unit circle or not.}
}
}
\description{
Provide an estimation of the value of pi using simple probabilities. The internal procedure uses C++ code.
}
\examples{
estimate_pi(B=5000, seed=10)
}
