% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{plot.pi}
\alias{plot.pi}
\title{Plotting of the points serving in the estimation of pi}
\usage{
\method{plot}{pi}(x)
}
\arguments{
\item{x}{an object of `pi` class containing the points that serve to the estimation of the pi value.}
}
\value{
A plot of the unit square and the unit circle emphasising the points used in the estimation.
}
\description{
Plotting procedure for an object of the `pi` class.
}
\examples{
plot(x)
}
